Completed transition of python function code to JS
Transitioned code from LPI.py to LPI.js.

*** NOTE ***
This code is almost completely untested!
It will be much easier to test with the well view (or with Python checking
the output LPF files).

Once tested, the LPFEncoder() class and all helper functions will be moved to
a separate file. LPI.py will also be deleted then.

Performance tests also need to be run.