Slight tweaks to well view. Trying to add feature #46
Slight aesthetic changes to well view and code cleanup.
Also trying to add ability to set plate view time point by clicking
on a time point in well view (#46). This mostly works now... that is,
`currentStep` is updated correctly, but the plate can't be updated to
reflect the new time because well view code doesn't have access to
the necessary updatePlate() or updateTime() functions. Not sure if
there's a good solution for this.