Added indicator for well index
To right of Row and Col indicators at top of visualization now
appears a Well # indicator for easier function programming.

Unfortunately (this was true before too, but more obvious now),
when selecting a well changes the number of indicated digits,
it changes the total text length, which moves the slider over a
bit. Not a very important issue. (#91)

Addresses #88. Could add a tooltip if we wanted, but this works.