Added max/min checking on func int writes
Added internal Function method that can check intensities to be
written to the intensity matrix. As of now, it just checks that
intensities are between 0 and 4095; if so, the value is returned.
Otherwise, the boundary value is returned and an error is thrown.

Where these errors are all going is still unknown...