Asynchronous errors, checked mult funcs on a channel
Closes #39. Error checking is now effectively complete; other features
(like #98 and #102) can be added later; they're pretty minimal and
moslty involve UI changes, not backend error stuff for the most part.

errorManager is the encapsulated way of dealing with all computational
errors - it catches everything LPFEncoder could throw. In the future,
it will cause UI changes in response to bad inputs and can reset the
Encoder's state (reinitialize).

Third-order (inter-function) error checking also complete, in the
simplest way possible: dynamic functions are not allowed to be written
to the same channel. Constant funcs can be applied, but are extraneous
and this functionality might be outlawed in the future as well. All
intensities are checked for validity before being written anyway.