Extended total run length, added Warning Class
To facilitate the occasional need to do very long runs (>1000min),
the input checking has been relaxed a bit on this parameter. To
help prevent the accidental creation of huge program files, a new
Warning class has been created, similar to the Error class, which
will warn users of his potentiality. Furthermore, the timestep is
automatically set to 10s instead of 1s for runs > 1000min.