Fixed popup binding, can minimize experiments
Experiment minimization now works, with the caveat that the button
can't be changed to it's minimized version until the CSS is updated
to reflect a new "minimized" class (#128 @felixekn).

Also fixed popup box binding, so now it's associated with a particular
WFG. Its method may posibly need more infomation about the WFG's index
or other properties to be able ti ID it properly in Plate, but that
will depend on how its chart is implemented later.