LEDs now turn off on finish
Closes #170. the plate.isSteadyState parameter is never reset from True --
if we care about determining when runs are steady-state runs, we should
fix this, otherwise, delete the variable.

Furthermore, Brian made a good point that currently the plate.offOnFinish
variable is only used during the creation of the LPF file, which means
that is never displayed on the simulation. A future commit will move this
computation to a more relevant part of Plate that ensures this setting is
able to be visualized.