DOwnload button now controlled by validation result
Defined global variable $scope.inputsValid (bool) which will hold the
overall result of the updateValidation() function. It is now used to
determine whether a new Plate object should be initialized when the inputs
have changed (no if the inputs are invalid, in order to avoid errors it
would throw), and whether the download button should be shown (no if
invalid).

Also fixed some issues with variables in updateDisplay() being used before
they were defined...