Added time validation, global tooltip settings
Added basic time validation, which works properly. For some reason, the
tooltip doesn't appear even though the box is correctly marked as invalid
and its title attribute is updated. Will make a separate issue for this,
as the validation is working properly and the angular variables are being
set correctly.

Also added global tooltip settings in a hail-mary effort to see if that
would help. Closes #196.