Implemented angular code for new error class for coloring invalid inputs
Added code to connect to the new `error` class, which when applied will
highlight the border of the input area red until the error is fixed.
Angular code is based on whether the title element is being set to nothing
(the empty string) or something (indicating an error on that field).