Updated Docs, reverted pairing select to disabled, not hidden
Updated the documentation with information on pairing Constant waveforms.
Also reverted the behavior of the pairing select input: it is no longer
hidden when dynamic waveforms are added to the experiment; instead, it's
just disabled. We felt that eliminating the clutter by hiding the select
was not a sufficiently good reason to make Iris' behavior more opaque to
the user. This way, it's clear exactly what's going on.

Fixes #215.