Attempt to address well view x-axis tick label crop issue
Refers to #230.

This is actually a surprisingly hard issue to solve robustly. For now,
I've just increased the maxmimum value for the x-axis, which provides a
little buffer room between the right edge of the plot and the last tick
mark.

This fixes the issue (#230) for ~80% of the total times I tested, and in
the cases where it fails, at least it's not truncating a tick that
corresponds to part of the program.

We can look into other fixes for this, but in the meantime I'm closing
this issue.

Fixes #230.