Made documentation consistent with Supp text
Improvements to documentation text & formatting, reconciling some rewrites
done for the paper's supplemental section.

A known new issue is that the staggered start image (nor any other image
for that matter) will not appear on the GitHub repo's README, as the
directory to the PNG has to be relative to the documentation folder for
the documentation.html file to get the image. May fix this by running
Pandoc on the correct dir, then changing the README to the other dir
before pushing (without re-running Pandoc). This may be annoying, we'll
see.