Changing step input parameters
Addresses #264: have reworked step input parameters in formServices.js and
in the HTML to accommodate new parameters and their error checking.

Validation seems to be working properly, as is the display.

Although the functional form was modified minimally to allow these
alternative inputs, the programs produced do not appear correct: the step
time appears to have no effect on the program. Need to dig deeper to
understand why.