Finalized docs for new step input format
Finalized documentation and double-checked the testing on the new step
input parameters & format. Added documentation, which is now simplified.
(Hopefully the LaTeX displays correctly; if not, it'll have to be
rewritten. I'll use Pandoc to compile this all later when we merge in the
inputStyles changes as well.)

Also added some details on the maximum program size and the changes in
time step that occur for long programs. Closes #263. Closes #265.