Added confirm dialog, selected class
Added dialog to confirm switching states, warning that all data will be
cleared. Only switches to new state if user selects 'OK'. The dialog box
is just the browser default, but it'll do for now. We can always make it
prettier later if we want (@felixekn).

Also added 'selected' class to the style transition buttons so that they
can be styled appropriately in the CSS.